Create package_rokit_index.json
This is the base .json format for the first "Live" version of the Rokit Smart Inventor boards.  The boards included are:
"Ardu-X100_v1"
"Ardu-X200_v1"
"SmartInventor-mega32_v1"
"SmartInventor-mega32_v2"

Do not change the format, title or location of this file/url.  To update, add additional versions instead of rewriting the .json.